refactor(storage)!: P7 pending store + rotation backup move
Physical move of InMemoryPendingStore from
crates/credential/src/pending_store_memory.rs to
crates/storage/src/credential/pending.rs per ADR-0029 §4 /
ADR-0032 (impls move, trait stays). PendingStateStore trait +
PendingStoreError remain in credential's pending_store.rs. A
body-identical shim is retained at
crates/credential/src/pending_store_memory.rs because executor.rs
internal tests reference it directly and cannot dev-dep on
nebula-storage — empirically confirmed P6.2 pattern.

Physical move of RotationBackup from
crates/credential/src/rotation/backup.rs to
crates/storage/src/credential/backup.rs. Feature-gated behind a
new \`rotation\` feature in storage that forwards to
\`nebula-credential/rotation\`; no backup users outside the
rotation subsystem.

PendingToken::generate() widened from pub(crate) to pub so the
canonical InMemoryPendingStore in storage can construct tokens;
documented as the intended cross-crate seam for pending-store
impls.

Migrated pending_lifecycle_tests.rs to
crates/storage/tests/credential_pending_lifecycle.rs. Extended
with dedicated tests pinning ADR-0029 §4 invariants 2-6 (TTL with
eviction, single-use, session-binding, SecretString compile-time
witness via OAuth2Pending, Zeroize trait-bound witness).
Invariant 1 (encryption at rest) is moot for the in-memory impl
and deferred to a future EncryptedPendingLayer per module doc and
ADR-0029 §4.1.

Also: added \`tracing\` to storage dependencies (needed by the
moved backup.rs restore-path instrumentation).

Ref: docs/adr/0029-storage-owns-credential-persistence.md §4
Ref: docs/adr/0032-credential-store-canonical-home.md
Ref: docs/superpowers/plans/2026-04-20-credential-cleanup-p6-p11.md §P7
