+++ crates/credential/Cargo.toml
@@ -14,10 +14,10 @@ async-trait = { workspace = true }
 # Tokio features (audited 2026-04-20, phase P3.1):
 #   `time`   — tokio::time items used across the crate:
 #              `timeout` in executor, resolver, refresh, rotation/validation;
-#              `sleep` in pending_store_memory;
+#              `sleep` in oauth2/flow (device-code polling);
 #              `Instant` + `sleep_until` in rotation/scheduler
 #   `sync`   — tokio::sync items used across the crate:
-#              `RwLock` in store_memory, pending_store_memory;
+#              `RwLock` in store_memory, pending_store_memory (dual-home shim);
 #              `Mutex` in resolver;
 #              `oneshot` + `Semaphore` in refresh (RefreshCoordinator)
 #   `macros` — #[tokio::test] (133 sites) and tokio::select! in rotation/scheduler
