+++ crates/credential/src/contract/pending.rs
@@ -84,7 +84,11 @@ impl std::fmt::Debug for PendingToken {
 
 impl PendingToken {
     /// Generates a new cryptographically random token.
-    pub(crate) fn generate() -> Self {
+    ///
+    /// Callable by pending-store impls that live outside this crate (see
+    /// `nebula_storage::credential::pending::InMemoryPendingStore` —
+    /// canonical home per ADR-0032 §7).
+    pub fn generate() -> Self {
         use rand::RngExt;
         let mut rng = rand::rng();
         let bytes: [u8; 32] = rng.random();
