+++ crates/credential/src/pending_store_memory.rs
@@ -1,7 +1,27 @@
-//! In-memory pending state store for testing and development.
+//! In-memory pending state store — **behaviour-identical shim** of the
+//! canonical impl at `nebula_storage::credential::pending::InMemoryPendingStore`
+//! (ADR-0029 §4 / ADR-0032 §7).
 //!
-//! Data is lost when the store is dropped. Use this in tests rather
-//! than mocking [`PendingStateStore`] directly.
+//! # Why a dual-home shim?
+//!
+//! Credential's own `executor.rs` `#[cfg(test)] mod tests` uses an
+//! `InMemoryPendingStore` to exercise `execute_resolve` / `execute_continue`
+//! against a real `PendingStateStore`. Those tests cannot dev-dep on
+//! `nebula-storage`: that dev-dep path produces a two-copies cargo
+//! resolution, which means the `PendingStateStore` trait bound on the
+//! generic executor no longer accepts the storage-side type
+//! (empirically confirmed in P6.2 with `InMemoryStore`).
+//!
+//! The fix is the same narrowly-scoped exception ADR-0032 §7 carved out
+//! for `store_memory`: keep a body-identical copy in credential for
+//! internal tests; production consumers and composition roots prefer the
+//! storage-side canonical home.
+//!
+//! If you are adding new production code, reach for
+//! `nebula_storage::credential::InMemoryPendingStore`. Only credential's
+//! internal `#[cfg(test)]` code should touch this module.
+//!
+//! Ref: `docs/adr/0032-credential-store-canonical-home.md` §7
 
 use std::{collections::HashMap, sync::Arc};
 
@@ -191,33 +211,19 @@ mod tests {
         }
     }
 
-    #[derive(Clone, Debug, Serialize, Deserialize, PartialEq)]
-    struct ShortLivedPending {
-        data: String,
-    }
-
-    impl Zeroize for ShortLivedPending {
-        fn zeroize(&mut self) {
-            self.data.zeroize();
-        }
-    }
-
-    impl PendingState for ShortLivedPending {
-        const KIND: &'static str = "short_lived";
-
-        fn expires_in(&self) -> Duration {
-            Duration::ZERO
-        }
-    }
-
     fn test_pending(data: &str) -> TestPending {
         TestPending {
             data: data.to_owned(),
         }
     }
 
+    // The full invariant / behaviour test matrix lives next to the
+    // canonical impl in `nebula_storage::credential::pending::tests` and
+    // in the integration test `credential_pending_lifecycle.rs`. Credential
+    // only retains a smoke test here — enough to catch shim drift without
+    // duplicating the matrix.
     #[tokio::test]
-    async fn put_and_consume_roundtrip() {
+    async fn shim_put_and_consume_roundtrip() {
         let store = InMemoryPendingStore::new();
         let pending = test_pending("hello");
 
@@ -232,161 +238,5 @@ mod tests {
             .unwrap();
 
         assert_eq!(result, pending);
-    }
-
-    #[tokio::test]
-    async fn consume_rejects_wrong_credential_kind_and_preserves_entry() {
-        let store = InMemoryPendingStore::new();
-        let token = store
-            .put("oauth2", "user_1", "sess_1", test_pending("x"))
-            .await
-            .unwrap();
-
-        let err = store
-            .consume::<TestPending>("api_key", &token, "user_1", "sess_1")
-            .await
-            .unwrap_err();
-        assert!(matches!(err, PendingStoreError::ValidationFailed { .. }));
-
-        // Entry must still be consumable by the legitimate caller — a
-        // wrong-kind probe must not destroy pending state.
-        let ok: TestPending = store
-            .consume("oauth2", &token, "user_1", "sess_1")
-            .await
-            .expect("legitimate consume should still succeed after bad probe");
-        assert_eq!(ok.data, "x");
-    }
-
-    #[tokio::test]
-    async fn consume_rejects_wrong_owner_and_preserves_entry() {
-        let store = InMemoryPendingStore::new();
-        let token = store
-            .put("oauth2", "user_1", "sess_1", test_pending("x"))
-            .await
-            .unwrap();
-
-        let err = store
-            .consume::<TestPending>("oauth2", &token, "user_2", "sess_1")
-            .await
-            .unwrap_err();
-        assert!(matches!(err, PendingStoreError::ValidationFailed { .. }));
-
-        let ok: TestPending = store
-            .consume("oauth2", &token, "user_1", "sess_1")
-            .await
-            .expect("legitimate consume should still succeed after bad probe");
-        assert_eq!(ok.data, "x");
-    }
-
-    #[tokio::test]
-    async fn consume_rejects_wrong_session_and_preserves_entry() {
-        let store = InMemoryPendingStore::new();
-        let token = store
-            .put("oauth2", "user_1", "sess_1", test_pending("x"))
-            .await
-            .unwrap();
-
-        let err = store
-            .consume::<TestPending>("oauth2", &token, "user_1", "sess_2")
-            .await
-            .unwrap_err();
-        assert!(matches!(err, PendingStoreError::ValidationFailed { .. }));
-
-        let ok: TestPending = store
-            .consume("oauth2", &token, "user_1", "sess_1")
-            .await
-            .expect("legitimate consume should still succeed after bad probe");
-        assert_eq!(ok.data, "x");
-    }
-
-    #[tokio::test]
-    async fn consume_deletes_entry() {
-        let store = InMemoryPendingStore::new();
-        let token = store
-            .put("oauth2", "user_1", "sess_1", test_pending("x"))
-            .await
-            .unwrap();
-
-        let _: TestPending = store
-            .consume("oauth2", &token, "user_1", "sess_1")
-            .await
-            .unwrap();
-
-        let err = store
-            .consume::<TestPending>("oauth2", &token, "user_1", "sess_1")
-            .await
-            .unwrap_err();
-
-        assert!(matches!(err, PendingStoreError::NotFound));
-    }
-
-    #[tokio::test]
-    async fn get_does_not_delete_entry() {
-        let store = InMemoryPendingStore::new();
-        let pending = test_pending("repeatable");
-        let token = store
-            .put("oauth2", "user_1", "sess_1", pending.clone())
-            .await
-            .unwrap();
-
-        let first: TestPending = store.get(&token).await.unwrap();
-        let second: TestPending = store.get(&token).await.unwrap();
-
-        assert_eq!(first, pending);
-        assert_eq!(second, pending);
-    }
-
-    #[tokio::test]
-    async fn expired_entry_returns_error_on_get() {
-        let store = InMemoryPendingStore::new();
-        let pending = ShortLivedPending {
-            data: "ephemeral".to_owned(),
-        };
-        let token = store
-            .put("oauth2", "user_1", "sess_1", pending)
-            .await
-            .unwrap();
-
-        tokio::time::sleep(Duration::from_millis(5)).await;
-
-        let err = store.get::<ShortLivedPending>(&token).await.unwrap_err();
-        assert!(matches!(err, PendingStoreError::Expired));
-        assert_eq!(store.entries.read().await.len(), 0);
-
-        // Once evicted, later reads are a clean miss instead of repeatedly
-        // returning Expired while retaining memory.
-        let second = store.get::<ShortLivedPending>(&token).await.unwrap_err();
-        assert!(matches!(second, PendingStoreError::NotFound));
-    }
-
-    #[tokio::test]
-    async fn expired_entry_returns_error_on_consume() {
-        let store = InMemoryPendingStore::new();
-        let pending = ShortLivedPending {
-            data: "ephemeral".to_owned(),
-        };
-        let token = store
-            .put("oauth2", "user_1", "sess_1", pending)
-            .await
-            .unwrap();
-
-        tokio::time::sleep(Duration::from_millis(5)).await;
-
-        let err = store
-            .consume::<ShortLivedPending>("oauth2", &token, "user_1", "sess_1")
-            .await
-            .unwrap_err();
-
-        assert!(matches!(err, PendingStoreError::Expired));
-    }
-
-    #[tokio::test]
-    async fn delete_is_idempotent() {
-        let store = InMemoryPendingStore::new();
-        let token = PendingToken::generate();
-
-        // Deleting a non-existent token should succeed.
-        store.delete(&token).await.unwrap();
-        store.delete(&token).await.unwrap();
     }
 }
