+++ crates/credential/tests/units/main.rs
@@ -6,7 +6,8 @@
 //! - Error handling and error types
 
 mod error_tests;
-mod pending_lifecycle_tests;
+// `pending_lifecycle_tests` moved to `crates/storage/tests/credential_pending_lifecycle.rs`
+// (ADR-0029 §4 / ADR-0032 — InMemoryPendingStore lives in nebula-storage).
 mod resolve_snapshot_tests;
 mod scheme_roundtrip_tests;
 mod thundering_herd_tests;
